add custom error instead of validaitonPipe

+++ src/user/user.controller.ts
@@ -1,4 +1,5 @@
 import {
+  BadRequestException,
   Body,
   Controller,
   Delete,
@@ -6,16 +7,19 @@ import {
   Param,
   Patch,
   Post,
-  ValidationPipe,
 } from '@nestjs/common';
 import { IUser, RegisterUserDto } from './userDto';
 import { UserService } from './user.service';
+import { throwCustomError } from 'src/utility/custom.error';
 
 @Controller('user')
 export class UserController {
   constructor(public readonly userService: UserService) {}
   @Post()
   createUser(@Body() body: RegisterUserDto) {
+    throwCustomError(body, 'logIn', 'string');
+    throwCustomError(body, 'password', 'string');
+    throwCustomError(body, 'email', 'string');
     this.userService.addUser(body);
   }
   @Get()
@@ -24,15 +28,27 @@ export class UserController {
   }
   @Get(':id')
   getUser(@Param('id') id: string) {
+    if (!id || typeof id !== 'string') {
+      throw new BadRequestException('Invalid request id');
+    }
     return this.userService.getUser(id);
   }
   @Patch(':id')
-  updateUser(@Param('id') id: string, @Body(ValidationPipe) Body: IUser) {
-    this.userService.updateUser(id, Body);
+  updateUser(@Param('id') id: string, @Body() body: IUser) {
+    throwCustomError(body, 'logIn', 'string');
+    throwCustomError(body, 'password', 'string');
+    throwCustomError(body, 'email', 'string');
+    if (!id || typeof id !== 'string') {
+      throw new BadRequestException('Invalid request id');
+    }
+    this.userService.updateUser(id, body);
     return null;
   }
   @Delete(':id')
   deleteUser(@Param('id') id: string) {
+    if (!id || typeof id !== 'string') {
+      throw new BadRequestException('Invalid request id');
+    }
     this.userService.deleteUser(id);
     return null;
   }
